+++ Source/SwiftLintCore/Extensions/Configuration+FileGraphSubtypes.swift
@@ -72,9 +72,9 @@ internal extension Configuration.FileGraph {
 
         private func read(at path: String) throws -> String {
             guard !path.isEmpty && FileManager.default.fileExists(atPath: path) else {
-                throw isInitialVertex ?
-                    Issue.initialFileNotFound(path: path) :
-                    Issue.genericWarning("File \(path) can't be found.")
+                throw isInitialVertex
+                    ? Issue.initialFileNotFound(path: path)
+                    : Issue.fileNotFound(path: path)
             }
 
             return try String(contentsOfFile: path, encoding: .utf8)